@@ -15,11 +15,11 @@ topics:
   - print
   - printing
   - report
-version: 5.12.0
+version: 5.13.0
 
 environment:
-  sdk: ">=3.0.0 <4.0.0"
-  flutter: ">=3.10.0"
+  sdk: ">=3.3.0 <4.0.0"
+  flutter: ">=3.19.0"
 
 dependencies:
   ffi: ">=1.1.0 <3.0.0"
@@ -29,14 +29,14 @@ dependencies:
     sdk: flutter
   http: ">=0.13.0 <2.0.0"
   image: ^4.0.02
-  js: ">=0.6.3 <1.0.0"
   meta: ">=1.3.0 <2.0.0"
   pdf: ^3.10.0
   pdf_widget_wrapper: '>=1.0.0 <2.0.0'
   plugin_platform_interface: ^2.1.0
+  web: ^0.5.1
 
 dev_dependencies:
-  flutter_lints: ^3.0.0
+  flutter_lints: ^4.0.0
   flutter_test:
     sdk: flutter
   mockito: ^5.3.2
