@@ -33,7 +33,7 @@ export 'page.dart' show PdfPreviewPageData;
 class PdfPreview extends StatefulWidget {
   /// Show a pdf document built on demand
   const PdfPreview({
-    Key? key,
+    super.key,
     required this.build,
     this.initialPageFormat,
     this.allowPrinting = true,
@@ -66,8 +66,7 @@ class PdfPreview extends StatefulWidget {
     this.actionBarTheme = const PdfActionBarTheme(),
     this.enableScrollToPage = false,
     this.onZoomChanged,
-  })  : _pagesBuilder = null,
-        super(key: key);
+  }) : _pagesBuilder = null;
 
   /// Build a custom layout.
   ///
@@ -93,7 +92,7 @@ class PdfPreview extends StatefulWidget {
   /// )
   /// ```
   const PdfPreview.builder({
-    Key? key,
+    super.key,
     required this.build,
     this.initialPageFormat,
     this.allowPrinting = true,
@@ -127,8 +126,7 @@ class PdfPreview extends StatefulWidget {
     required CustomPdfPagesBuilder pagesBuilder,
     this.enableScrollToPage = false,
     this.onZoomChanged,
-  })  : _pagesBuilder = pagesBuilder,
-        super(key: key);
+  }) : _pagesBuilder = pagesBuilder;
 
   static const _defaultPageFormats = <String, PdfPageFormat>{
     'A4': PdfPageFormat.a4,
@@ -341,29 +339,40 @@ class PdfPreviewState extends State<PdfPreview> {
     final actions = <Widget>[];
 
     if (widget.useActions && widget.allowPrinting && info?.canPrint == true) {
-      actions.add(PdfPrintAction(
-        jobName: widget.pdfFileName,
-        dynamicLayout: widget.dynamicLayout,
-        onPrinted:
-            widget.onPrinted == null ? null : () => widget.onPrinted!(context),
-        onPrintError: widget.onPrintError == null
-            ? null
-            : (dynamic error) => widget.onPrintError!(context, error),
-      ));
+      actions.add(
+        PdfPrintAction(
+          jobName: widget.pdfFileName,
+          dynamicLayout: widget.dynamicLayout,
+          onPrinted: widget.onPrinted == null
+              ? null
+              : () => widget.onPrinted!(context),
+          onPrintError: widget.onPrintError == null
+              ? null
+              : (dynamic error) => widget.onPrintError!(context, error),
+        ),
+      );
     }
 
     if (widget.useActions && widget.allowSharing && info?.canShare == true) {
-      actions.add(PdfShareAction(
-        filename: widget.pdfFileName,
-        onShared:
-            widget.onPrinted == null ? null : () => widget.onPrinted!(context),
-      ));
+      actions.add(
+        PdfShareAction(
+          filename: widget.pdfFileName,
+          onShared: widget.onPrinted == null
+              ? null
+              : () => widget.onPrinted!(context),
+          subject: widget.shareActionExtraSubject,
+          emails: widget.shareActionExtraEmails,
+          body: widget.shareActionExtraBody,
+        ),
+      );
     }
 
     if (widget.useActions && widget.canChangePageFormat) {
-      actions.add(PdfPageFormatAction(
-        pageFormats: widget.pageFormats,
-      ));
+      actions.add(
+        PdfPageFormatAction(
+          pageFormats: widget.pageFormats,
+        ),
+      );
     }
 
     if (widget.useActions && widget.canChangeOrientation) {
@@ -398,27 +407,29 @@ class PdfPreviewState extends State<PdfPreview> {
         mainAxisAlignment: MainAxisAlignment.center,
         children: <Widget>[
           Expanded(
-            child: Builder(builder: (context) {
-              final controller = PdfPreviewController.listen(context);
-              return PdfPreviewCustom(
-                key: previewWidget,
-                build: controller.buildDocument,
-                loadingWidget: widget.loadingWidget,
-                maxPageWidth: widget.maxPageWidth,
-                onError: widget.onError,
-                padding: widget.padding,
-                pageFormat: controller.pageFormat,
-                pages: widget.pages,
-                pdfPreviewPageDecoration: widget.pdfPreviewPageDecoration,
-                previewPageMargin: widget.previewPageMargin,
-                scrollViewDecoration: widget.scrollViewDecoration,
-                shouldRepaint: widget.shouldRepaint,
-                pagesBuilder: widget._pagesBuilder,
-                dpi: widget.dpi,
-                enableScrollToPage: widget.enableScrollToPage,
-                onZoomChanged: widget.onZoomChanged,
-              );
-            }),
+            child: Builder(
+              builder: (context) {
+                final controller = PdfPreviewController.listen(context);
+                return PdfPreviewCustom(
+                  key: previewWidget,
+                  build: controller.buildDocument,
+                  loadingWidget: widget.loadingWidget,
+                  maxPageWidth: widget.maxPageWidth,
+                  onError: widget.onError,
+                  padding: widget.padding,
+                  pageFormat: controller.pageFormat,
+                  pages: widget.pages,
+                  pdfPreviewPageDecoration: widget.pdfPreviewPageDecoration,
+                  previewPageMargin: widget.previewPageMargin,
+                  scrollViewDecoration: widget.scrollViewDecoration,
+                  shouldRepaint: widget.shouldRepaint,
+                  pagesBuilder: widget._pagesBuilder,
+                  dpi: widget.dpi,
+                  enableScrollToPage: widget.enableScrollToPage,
+                  onZoomChanged: widget.onZoomChanged,
+                );
+              },
+            ),
           ),
           if (actions.isNotEmpty)
             IconTheme.merge(
